+++ Button.js
@@ -1,0 +1,18 @@
+export class Button {
+  constructor({ title, value, type }) {
+    this.title = title;
+    this.value = value;
+    this.type = type;
+
+    return this.render();
+  }
+
+  render() {
+    const button = document.createElement('button');
+    button.innerHTML = this.title;
+    button.setAttribute('data-value', this.value);
+    button.setAttribute('data-type', this.type);
+    button.classList.add('button');
+    return button;
+  }
+}
