+++ index.css
@@ -1,38 +1,90 @@
 .wrapper {
   display: flex;
   flex-direction: column;
+  row-gap: 3vh;
   align-items: center;
   justify-items: center;
-  width: 60vw;
+  width: 40vw;
   margin: 0 auto;
+  padding-top: 10vh;
 }
 
 .input {
   height: 10vh;
-  width: 60%;
+  width: 100%;
   font-size: 56px;
   text-align: end;
+  padding: 0;
+}
+
+.input:focus {
+  outline: none;
 }
 
 .numPad {
   display: flex;
-  flex-wrap: wrap-reverse;
+  width: 100%;
+  column-gap: 1vh;
+  justify-content: center;
   align-items: center;
   justify-items: space-around;
-  width: 60%;
+}
+
+.numbersWrapper {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  align-items: center;
+  flex-grow: 1;
+  grid-gap: 1vh;
+}
+.operatorsWrapper {
+  display: grid;
+  grid-template-rows: repeat(4, 1fr);
+  grid-template-columns: repeat(2, 1fr);
+  align-items: center;
+  flex-grow: 1;
+  max-width: 35%;
+  grid-gap: 1vh;
 }
 
 .operatorPud {
   display: flex;
-  flex-wrap: wrap-reverse;
+  flex-wrap: wrap;
+
   align-items: center;
   justify-items: space-around;
   width: 100%;
 }
 .button {
-  margin: 1vw;
-  min-height: 5vh;
-  min-width: 5vh;
+  min-height: 7vh;
   font-size: 30px;
   cursor: pointer;
+  border: none;
+  border-radius: 5px;
+}
+
+.button:hover {
+  /* background-color: rgb(162, 161, 161); */
+  box-shadow: 5px 5px 15px rgb(184, 184, 184);
+}
+
+.numButton {
+  background-color: light-gray;
+}
+.operatorButton {
+  background-color: black;
+  color: white;
+}
+.clearButton {
+  background-color: rgb(74, 9, 5);
+  color: white;
+}
+.resultButton {
+  grid-column: 2;
+  grid-row: 3 / 5;
+  height: 100%;
+}
+.funcButton {
+  background-color: rgb(15, 5, 70);
+  color: white;
 }
